refactor(mbk/leases): split signed_lease_service.py (1,647 LOC) into focused modules
Splits the monolithic signed_lease_service.py into six focused modules:

  _lease_helpers.py         (255 LOC) — exceptions, constants, shared private helpers
  lease_lifecycle_service.py (306 LOC) — create / list / get / update / soft-delete
  lease_pdf_service.py      (501 LOC) — generate_lease / add_templates_and_generate
  lease_prefill_service.py  (144 LOC) — prefill_addendum_placeholders
  lease_import_service.py   (258 LOC) — import_signed_lease + kind-inference heuristics
  lease_attachment_service.py (251 LOC) — upload / list / update / delete attachments

The original signed_lease_service.py becomes a 109-line re-export shim so all
existing callers (API routes, tests) continue to work without import changes.

Test updates: patch paths in test_lease_add_template.py updated from
signed_lease_service.* to the focused module that now owns unit_of_work /
get_storage / render_md. test_lease_attachment_kind and test_lease_import_heuristic
updated to import directly from the focused modules. All 2620 tests pass.
