+++ apps/mybookkeeper/backend/app/services/leases/signed_lease_service.py
@@ -1,1647 +1,109 @@
-"""Service layer for signed leases.
+"""Signed lease service — thin re-export module.
 
-Handles the lease lifecycle: draft → generated → sent → signed → active.
+All implementation has been split into focused modules. This file re-exports
+every public name so existing callers (API routes, workers, tests) continue
+to work without changes.
 
-Pipeline for ``generate``:
-    load template files → run placeholder substitution per file →
-    upload rendered output to MinIO under ``signed-leases/<lease_id>/`` →
-    create ``signed_lease_attachments`` rows with kind=rendered_original →
-    transition status to ``generated`` and stamp ``generated_at`` →
-    re-fetch detail with attachments and presigned URLs.
+Implementation lives in:
+  _lease_helpers.py         — exceptions, constants, private shared helpers
+  lease_lifecycle_service.py — create / list / get / update / soft-delete
+  lease_pdf_service.py      — generate / add_templates_and_generate
+  lease_prefill_service.py  — prefill_addendum_placeholders
+  lease_import_service.py   — import + kind-inference heuristics
+  lease_attachment_service.py — upload / list / update / delete attachments
 """
 from __future__ import annotations
 
-import datetime as _dt
-import logging
-import uuid
-from typing import Any
-
-from app.core.lease_enums import LEASE_ATTACHMENT_KINDS, SIGNED_LEASE_STATUSES
-from app.core.storage import get_storage
-from app.db.session import unit_of_work
-from app.repositories.applicants import applicant_repo
-from app.repositories.inquiries.inquiry_repo import get_by_applicant_inquiry_id
-from app.repositories.leases import (
-    lease_template_file_repo,
-    lease_template_placeholder_repo,
-    lease_template_repo,
-    signed_lease_attachment_repo,
-    signed_lease_repo,
-    signed_lease_template_repo,
-)
-from app.repositories.listings import listing_repo
-from app.repositories.properties import property_repo
-from app.repositories.user import user_repo
-from app.models.applicants.applicant import Applicant
-from app.models.inquiries.inquiry import Inquiry
-from app.models.properties.property import Property
-from app.models.user.user import User
-from app.schemas.leases.signed_lease_attachment_response import (
-    SignedLeaseAttachmentResponse,
-)
-from app.schemas.leases.signed_lease_list_response import (
-    SignedLeaseListResponse,
-)
-from app.schemas.leases.signed_lease_response import SignedLeaseResponse
-from app.schemas.leases.signed_lease_summary import SignedLeaseSummary
-from app.schemas.leases.signed_lease_template_link import SignedLeaseTemplateLink
-from app.schemas.leases.signed_lease_template_prefill_response import (
-    SignedLeaseTemplatePrefillItem,
-    SignedLeaseTemplatePrefillResponse,
-)
-from app.services.leases.attachment_response_builder import (
-    attach_presigned_urls_to_attachments,
-)
-from app.services.leases.computed import ComputedExprError, evaluate
-from app.services.leases.default_source_resolver import resolve_default_source
-from app.services.leases.renderer import (
-    render_docx_bytes_to_pdf,
-    render_md,
-    render_pdf_from_text,
-)
-from app.services.leases.lease_template_service import (
-    DOCX_MIME,
-    TemplateNotFoundError,
+# Exceptions (from helpers — shared across all modules)
+from app.services.leases._lease_helpers import (
+    AttachmentNotFoundError,
+    AttachmentTooLargeError,
+    AttachmentTypeRejectedError,
+    ALLOWED_ATTACHMENT_MIME_TYPES,
+    CannotEditValuesError,
+    InvalidAttachmentKindError,
+    InvalidStatusTransitionError,
+    MissingRequiredValuesError,
+    SignedLeaseNotFoundError,
+    StorageNotConfiguredError,
 )
 
-logger = logging.getLogger(__name__)
+# Lifecycle
+from app.services.leases.lease_lifecycle_service import (
+    create_lease,
+    get_lease,
+    list_leases,
+    soft_delete_lease,
+    update_lease,
+)
+
+# PDF generation
+from app.services.leases.lease_pdf_service import (
+    ImportedLeaseTemplateError,
+    TemplatesAlreadyLinkedError,
+    add_templates_and_generate,
+    generate_lease,
+    render_md_text_to_pdf_or_keep,
+    should_auto_email_after_generate,
+)
+
+# Placeholder pre-fill
+from app.services.leases.lease_prefill_service import prefill_addendum_placeholders
+
+# Import
+from app.services.leases.lease_import_service import (
+    ApplicantNotFoundError,
+    ListingNotFoundError,
+    import_signed_lease,
+    infer_kind_from_filename,
+    infer_kinds_for_files,
+)
 
-
-# ---------------------------------------------------------------------------
-# Exceptions
-# ---------------------------------------------------------------------------
-
-class SignedLeaseNotFoundError(LookupError):
-    pass
-
-
-class AttachmentNotFoundError(LookupError):
-    pass
-
-
-class StorageNotConfiguredError(RuntimeError):
-    pass
-
-
-class MissingRequiredValuesError(ValueError):
-    pass
-
-
-class InvalidStatusTransitionError(ValueError):
-    pass
-
-
-class CannotEditValuesError(ValueError):
-    """``values`` can only be edited while status=draft."""
-
-
-class AttachmentTooLargeError(ValueError):
-    pass
-
-
-class AttachmentTypeRejectedError(ValueError):
-    pass
-
-
-class InvalidAttachmentKindError(ValueError):
-    pass
-
-
-# ---------------------------------------------------------------------------
-# Allowlist for signed-lease attachment uploads.
-# ---------------------------------------------------------------------------
-
-ALLOWED_ATTACHMENT_MIME_TYPES: frozenset[str] = frozenset({
-    "application/pdf",
-    "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-    "image/jpeg",
-    "image/png",
-    "image/webp",
-})
-
-# Status transitions allowed by the host. Backwards transitions are disallowed
-# so an "active" lease can't be flipped back to "draft" without a tech reset.
-_ALLOWED_TRANSITIONS: dict[str, set[str]] = {
-    "draft": {"draft", "generated"},
-    "generated": {"generated", "sent", "signed"},
-    "sent": {"sent", "signed"},
-    "signed": {"signed", "active"},
-    "active": {"active", "ended", "terminated"},
-    "ended": {"ended"},
-    "terminated": {"terminated"},
-}
-
-
-def infer_kind_from_filename(filename: str) -> str:
-    """Infer the attachment kind from a filename using pattern matching.
-
-    Order of evaluation (case-insensitive):
-    1. "move-in inspection" / "move in inspection" → move_in_inspection
-    2. "move-out inspection" / "move out inspection" → move_out_inspection
-    3. "lease agreement" / "master lease" / "rental agreement" → signed_lease
-    4. "inspection" (without "move") → move_in_inspection (default to in if ambiguous)
-    5. "insurance" → insurance_proof
-    6. Everything else → signed_addendum
-    """
-    lower = filename.lower()
-
-    if "move-in inspection" in lower or "move in inspection" in lower:
-        return "move_in_inspection"
-    if "move-out inspection" in lower or "move out inspection" in lower:
-        return "move_out_inspection"
-    if "lease agreement" in lower or "master lease" in lower or "rental agreement" in lower:
-        return "signed_lease"
-    if "inspection" in lower:
-        return "move_in_inspection"
-    if "insurance" in lower:
-        return "insurance_proof"
-    return "signed_addendum"
-
-
-def infer_kinds_for_files(filenames: list[str]) -> list[str]:
-    """Infer a kind for each filename in a batch.
-
-    Applies ``infer_kind_from_filename`` to each file. If none of the
-    inferred kinds is ``signed_lease``, the first file is promoted to
-    ``signed_lease`` as a last-resort fallback so every batch has at
-    least one main lease.
-    """
-    kinds = [infer_kind_from_filename(name) for name in filenames]
-    if "signed_lease" not in kinds and filenames:
-        kinds[0] = "signed_lease"
-    return kinds
-
-
-def _validate_status_transition(current: str, target: str) -> None:
-    if target == current:
-        return
-    if target not in SIGNED_LEASE_STATUSES:
-        raise InvalidStatusTransitionError(f"Unknown status: {target}")
-    if target not in _ALLOWED_TRANSITIONS.get(current, {current}):
-        raise InvalidStatusTransitionError(
-            f"Cannot move from '{current}' to '{target}'"
-        )
-
-
-def _build_summary(
-    lease,
-    applicant_names: dict[uuid.UUID, str | None] | None = None,
-    template_ids: list[uuid.UUID] | None = None,
-) -> SignedLeaseSummary:
-    summary = SignedLeaseSummary.model_validate(lease)
-    updates: dict[str, Any] = {"template_ids": template_ids or []}
-    if applicant_names is not None:
-        updates["applicant_legal_name"] = applicant_names.get(lease.applicant_id)
-    return summary.model_copy(update=updates)
-
-
-def _attachment_responses(rows) -> list[SignedLeaseAttachmentResponse]:
-    return attach_presigned_urls_to_attachments(
-        [SignedLeaseAttachmentResponse.model_validate(r) for r in rows],
-    )
-
-
-def _to_detail(lease, attachments, template_links: list[SignedLeaseTemplateLink]) -> SignedLeaseResponse:
-    return SignedLeaseResponse(
-        id=lease.id,
-        user_id=lease.user_id,
-        organization_id=lease.organization_id,
-        templates=template_links,
-        applicant_id=lease.applicant_id,
-        listing_id=lease.listing_id,
-        kind=lease.kind,
-        values=dict(lease.values or {}),
-        status=lease.status,
-        starts_on=lease.starts_on,
-        ends_on=lease.ends_on,
-        notes=lease.notes,
-        generated_at=lease.generated_at,
-        sent_at=lease.sent_at,
-        signed_at=lease.signed_at,
-        ended_at=lease.ended_at,
-        auto_email_tenant=lease.auto_email_tenant,
-        last_emailed_to_tenant_at=lease.last_emailed_to_tenant_at,
-        created_at=lease.created_at,
-        updated_at=lease.updated_at,
-        attachments=_attachment_responses(attachments),
-    )
-
-
-async def _resolve_template_links(
-    db,
-    *,
-    lease_id: uuid.UUID,
-) -> list[SignedLeaseTemplateLink]:
-    """Resolve the ordered list of template links (id + name + version) for a lease."""
-    join_rows = await signed_lease_template_repo.list_for_lease(db, lease_id=lease_id)
-    if not join_rows:
-        return []
-    template_ids = [r.template_id for r in join_rows]
-    # Bulk-load template metadata via the repo (one IN-list query).
-    templates = await lease_template_repo.list_by_ids(
-        db, template_ids=template_ids,
-    )
-    templates_by_id = {t.id: t for t in templates}
-    links: list[SignedLeaseTemplateLink] = []
-    for r in join_rows:
-        template = templates_by_id.get(r.template_id)
-        if template is None:
-            # Defensive — RESTRICT FK should prevent this, but skip gracefully.
-            continue
-        links.append(
-            SignedLeaseTemplateLink(
-                id=template.id,
-                name=template.name,
-                version=template.version,
-                display_order=r.display_order,
-            )
-        )
-    return links
-
-
-def _denormalise_dates(values: dict[str, Any]) -> tuple[_dt.date | None, _dt.date | None]:
-    """Pull ``starts_on`` and ``ends_on`` from a values dict if present."""
-    def _coerce(v: Any) -> _dt.date | None:
-        if v is None:
-            return None
-        if isinstance(v, _dt.date) and not isinstance(v, _dt.datetime):
-            return v
-        if isinstance(v, str):
-            try:
-                return _dt.date.fromisoformat(v)
-            except ValueError:
-                return None
-        return None
-
-    candidates_start = ("MOVE-IN DATE", "MOVE_IN_DATE", "MOVE IN DATE")
-    candidates_end = ("MOVE-OUT DATE", "MOVE_OUT_DATE", "MOVE OUT DATE")
-    starts = next((_coerce(values[k]) for k in candidates_start if k in values), None)
-    ends = next((_coerce(values[k]) for k in candidates_end if k in values), None)
-    return starts, ends
-
-
-async def _load_resolution_context(
-    db,
-    *,
-    lease,
-    organization_id: uuid.UUID,
-    user_id: uuid.UUID,
-) -> tuple[Applicant | None, Inquiry | None, Property | None, User | None]:
-    """Load applicant + inquiry + property + host user for placeholder resolution.
-
-    Used by ``add_templates_and_generate`` and ``prefill_addendum_placeholders``
-    to populate ``default_source`` resolution against live ORM rows. All four
-    return values may be ``None`` — the resolver is expected to tolerate
-    missing context (e.g. an imported lease with no listing has no property).
-    """
-    applicant = await applicant_repo.get(
-        db,
-        applicant_id=lease.applicant_id,
-        organization_id=organization_id,
-        user_id=user_id,
-    )
-    inquiry = None
-    if applicant is not None and applicant.inquiry_id is not None:
-        inquiry = await get_by_applicant_inquiry_id(
-            db,
-            inquiry_id=applicant.inquiry_id,
-            organization_id=organization_id,
-            user_id=user_id,
-        )
-    property_record = None
-    if lease.listing_id is not None:
-        listing = await listing_repo.get_by_id(
-            db, lease.listing_id, organization_id,
-        )
-        if listing is not None and listing.property_id is not None:
-            property_record = await property_repo.get_by_id(
-                db, listing.property_id, organization_id,
-            )
-    user_record = await user_repo.get_by_id(db, user_id)
-    return applicant, inquiry, property_record, user_record
-
-
-# ---------------------------------------------------------------------------
-# Create a draft lease
-# ---------------------------------------------------------------------------
-
-async def create_lease(
-    *,
-    user_id: uuid.UUID,
-    organization_id: uuid.UUID,
-    template_ids: list[uuid.UUID],
-    applicant_id: uuid.UUID,
-    listing_id: uuid.UUID | None,
-    values: dict[str, Any],
-) -> SignedLeaseResponse:
-    """Create a draft signed lease from one or more templates.
-
-    Validates required placeholders across the union of all selected
-    templates. Persists ONE ``signed_leases`` row plus N rows in
-    ``signed_lease_templates`` (one per template, ordered by
-    ``display_order`` matching the host's pick order).
-    """
-    if not template_ids:
-        raise TemplateNotFoundError("At least one template_id is required")
-
-    async with unit_of_work() as db:
-        # Validate every template is in scope and load placeholders for each.
-        # Union the required-placeholder set across all templates so the
-        # generate form's missing-values check covers every contributing doc.
-        seen_keys: set[str] = set()
-        merged_required: list = []
-        for tid in template_ids:
-            template = await lease_template_repo.get(
-                db,
-                template_id=tid,
-                user_id=user_id,
-                organization_id=organization_id,
-            )
-            if template is None:
-                raise TemplateNotFoundError(f"Template {tid} not found")
-            placeholders = await lease_template_placeholder_repo.list_for_template(
-                db, template_id=tid,
-            )
-            for p in placeholders:
-                if p.key in seen_keys:
-                    continue
-                seen_keys.add(p.key)
-                merged_required.append(p)
-
-        # Validate required placeholders against the merged set.
-        missing: list[str] = []
-        for p in merged_required:
-            if not p.required:
-                continue
-            if p.input_type == "computed":
-                continue
-            if p.input_type == "signature":
-                # Filled at signing time, not now.
-                continue
-            if values.get(p.key) in (None, ""):
-                missing.append(p.key)
-        if missing:
-            raise MissingRequiredValuesError(
-                f"Missing required values: {', '.join(missing)}"
-            )
-
-        starts, ends = _denormalise_dates(values)
-
-        lease = await signed_lease_repo.create(
-            db,
-            user_id=user_id,
-            organization_id=organization_id,
-            applicant_id=applicant_id,
-            listing_id=listing_id,
-            values=values,
-            starts_on=starts,
-            ends_on=ends,
-            status="draft",
-            kind="generated",
-        )
-        # Persist the ordered template links.
-        for order, tid in enumerate(template_ids):
-            await signed_lease_template_repo.create(
-                db,
-                lease_id=lease.id,
-                template_id=tid,
-                display_order=order,
-            )
-
-        attachments = await signed_lease_attachment_repo.list_by_lease(db, lease.id)
-        template_links = await _resolve_template_links(db, lease_id=lease.id)
-    return _to_detail(lease, attachments, template_links)
-
-
-# ---------------------------------------------------------------------------
-# List + get
-# ---------------------------------------------------------------------------
-
-async def list_leases(
-    *,
-    user_id: uuid.UUID,
-    organization_id: uuid.UUID,
-    applicant_id: uuid.UUID | None = None,
-    listing_id: uuid.UUID | None = None,
-    status: str | None = None,
-    starts_after: _dt.date | None = None,
-    starts_before: _dt.date | None = None,
-    limit: int = 50,
-    offset: int = 0,
-) -> SignedLeaseListResponse:
-    async with unit_of_work() as db:
-        rows = await signed_lease_repo.list_for_tenant(
-            db,
-            user_id=user_id,
-            organization_id=organization_id,
-            applicant_id=applicant_id,
-            listing_id=listing_id,
-            status=status,
-            starts_after=starts_after,
-            starts_before=starts_before,
-            limit=limit,
-            offset=offset,
-        )
-        total = await signed_lease_repo.count_for_tenant(
-            db,
-            user_id=user_id,
-            organization_id=organization_id,
-            applicant_id=applicant_id,
-            listing_id=listing_id,
-            status=status,
-        )
-
-        # Bulk-load applicant names for the tenant name column on the list page.
-        # Single IN-list query; the EncryptedString TypeDecorator decrypts
-        # transparently on load.
-        applicant_ids = [aid for aid in {r.applicant_id for r in rows} if aid is not None]
-        applicants = await applicant_repo.list_by_ids(
-            db,
-            organization_id=organization_id,
-            user_id=user_id,
-            applicant_ids=applicant_ids,
-        )
-        applicant_names: dict[uuid.UUID, str | None] = {
-            a.id: a.legal_name for a in applicants
-        }
-
-        # Bulk-load template IDs per lease (ordered) — one IN-list query.
-        template_ids_by_lease = await signed_lease_template_repo.list_template_ids_for_leases(
-            db, lease_ids=[r.id for r in rows],
-        )
-
-    items = [
-        _build_summary(r, applicant_names, template_ids_by_lease.get(r.id, []))
-        for r in rows
-    ]
-    return SignedLeaseListResponse(
-        items=items, total=total, has_more=(offset + len(items)) < total,
-    )
-
-
-async def get_lease(
-    *,
-    user_id: uuid.UUID,
-    organization_id: uuid.UUID,
-    lease_id: uuid.UUID,
-) -> SignedLeaseResponse:
-    async with unit_of_work() as db:
-        lease = await signed_lease_repo.get(
-            db,
-            lease_id=lease_id,
-            user_id=user_id,
-            organization_id=organization_id,
-        )
-        if lease is None:
-            raise SignedLeaseNotFoundError(f"Lease {lease_id} not found")
-        attachments = await signed_lease_attachment_repo.list_by_lease(
-            db, lease.id,
-        )
-        template_links = await _resolve_template_links(db, lease_id=lease.id)
-    return _to_detail(lease, attachments, template_links)
-
-
-# ---------------------------------------------------------------------------
-# Update
-# ---------------------------------------------------------------------------
-
-async def update_lease(
-    *,
-    user_id: uuid.UUID,
-    organization_id: uuid.UUID,
-    lease_id: uuid.UUID,
-    notes: str | None,
-    status: str | None,
-    values: dict[str, Any] | None,
-    auto_email_tenant: bool | None = None,
-) -> SignedLeaseResponse:
-    async with unit_of_work() as db:
-        lease = await signed_lease_repo.get(
-            db,
-            lease_id=lease_id,
-            user_id=user_id,
-            organization_id=organization_id,
-        )
-        if lease is None:
-            raise SignedLeaseNotFoundError(f"Lease {lease_id} not found")
-
-        fields: dict[str, Any] = {}
-        if notes is not None:
-            fields["notes"] = notes
-
-        if status is not None:
-            _validate_status_transition(lease.status, status)
-            fields["status"] = status
-            now = _dt.datetime.now(_dt.timezone.utc)
-            if status == "sent" and lease.sent_at is None:
-                fields["sent_at"] = now
-            if status == "signed" and lease.signed_at is None:
-                fields["signed_at"] = now
-            if status in ("ended", "terminated") and lease.ended_at is None:
-                fields["ended_at"] = now
-
-        if auto_email_tenant is not None:
-            fields["auto_email_tenant"] = auto_email_tenant
-
-        if values is not None:
-            if lease.status != "draft":
-                raise CannotEditValuesError(
-                    "Values can only be edited while the lease is a draft",
-                )
-            fields["values"] = values
-            starts, ends = _denormalise_dates(values)
-            fields["starts_on"] = starts
-            fields["ends_on"] = ends
-
-        if fields:
-            await signed_lease_repo.update_lease(
-                db,
-                lease_id=lease_id,
-                user_id=user_id,
-                organization_id=organization_id,
-                fields=fields,
-            )
-
-        attachments = await signed_lease_attachment_repo.list_by_lease(
-            db, lease_id,
-        )
-        # Re-load lease so we return updated values.
-        lease = await signed_lease_repo.get(
-            db,
-            lease_id=lease_id,
-            user_id=user_id,
-            organization_id=organization_id,
-        )
-        template_links = await _resolve_template_links(db, lease_id=lease_id)
-    return _to_detail(lease, attachments, template_links)
-
-
-# ---------------------------------------------------------------------------
-# Generate (renders the template files into MinIO + creates attachments)
-# ---------------------------------------------------------------------------
-
-class TemplatesAlreadyLinkedError(ValueError):
-    """One or more template_ids are already linked to this lease."""
-
-    def __init__(self, duplicate_ids: list[uuid.UUID]) -> None:
-        self.duplicate_ids = duplicate_ids
-        super().__init__(f"Templates already linked: {duplicate_ids}")
-
-
-class ImportedLeaseTemplateError(ValueError):
-    """Cannot add templates to an imported lease."""
-
-
-# ---------------------------------------------------------------------------
-# Add templates to an existing generated lease + render only the new files
-# ---------------------------------------------------------------------------
-
-async def add_templates_and_generate(
-    *,
-    user_id: uuid.UUID,
-    organization_id: uuid.UUID,
-    lease_id: uuid.UUID,
-    template_ids: list[uuid.UUID],
-    values_override: dict[str, str] | None = None,
-) -> SignedLeaseResponse:
-    """Link additional templates to an existing lease and render only their files.
-
-    Validates:
-    - All ``template_ids`` exist and belong to the same org as the lease.
-    - None are already linked to the lease (returns the duplicates in the error).
-
-    Both generated and imported leases are supported. For imported leases
-    ``lease.values`` is empty by construction (the original was uploaded as a
-    PDF) — the caller may pass ``values_override`` to provide values for the
-    addendum's placeholders, and the resolver will additionally pre-fill any
-    placeholder whose ``default_source`` resolves against the parent lease,
-    its applicant, the linked property, or the host user.
-
-    Then, for each new template:
-    - Inserts a row in ``signed_lease_templates`` with display_order = max + N.
-    - Downloads each template file, runs placeholder substitution, and uploads
-      the rendered output as a new ``SignedLeaseAttachment``
-      (kind=``rendered_original``).
-    - Does NOT touch existing attachments.
-
-    Partial success: if one template's render fails the DB row and MinIO
-    objects for that template are rolled back; others succeed. Failures are
-    logged and returned in the response's ``failed_template_ids`` field.
-    """
-    storage = get_storage()
-    if storage is None:
-        raise StorageNotConfiguredError("Object storage is not configured")
-
-    async with unit_of_work() as db:
-        lease = await signed_lease_repo.get(
-            db,
-            lease_id=lease_id,
-            user_id=user_id,
-            organization_id=organization_id,
-        )
-        if lease is None:
-            raise SignedLeaseNotFoundError(f"Lease {lease_id} not found")
-
-        # Validate every template is in scope.
-        for tid in template_ids:
-            template = await lease_template_repo.get(
-                db,
-                template_id=tid,
-                user_id=user_id,
-                organization_id=organization_id,
-            )
-            if template is None:
-                raise TemplateNotFoundError(f"Template {tid} not found")
-
-        # Already-linked templates being re-attached are treated as a
-        # regenerate request: their existing rendered files will be
-        # replaced rather than 409-rejected. Genuine new templates get
-        # join rows; duplicates skip the join row create (already linked).
-        existing_rows = await signed_lease_template_repo.list_for_lease(
-            db, lease_id=lease_id,
-        )
-        existing_ids = {r.template_id for r in existing_rows}
-        new_template_ids = [tid for tid in template_ids if tid not in existing_ids]
-        is_regenerate = any(tid in existing_ids for tid in template_ids)
-
-        # Compute the starting display_order for the new links.
-        max_order = await signed_lease_template_repo.max_display_order_for_lease(
-            db, lease_id=lease_id,
-        )
-
-        # Load files + placeholders. On regenerate (any template_id matches an
-        # existing link), re-render ALL linked templates' files so the
-        # generated documents stay internally consistent with lease.values.
-        # On a pure add (no duplicates), only render the new templates' files.
-        files_to_render: list = []         # (template_file_row, template_id)
-        placeholders: list = []
-        seen_placeholder_keys: set[str] = set()
-        seen_file_template_ids: set[uuid.UUID] = set()
-
-        for r in existing_rows:
-            for p in await lease_template_placeholder_repo.list_for_template(
-                db, template_id=r.template_id,
-            ):
-                if p.key not in seen_placeholder_keys:
-                    seen_placeholder_keys.add(p.key)
-                    placeholders.append(p)
-            if is_regenerate and r.template_id not in seen_file_template_ids:
-                seen_file_template_ids.add(r.template_id)
-                tpl_files = await lease_template_file_repo.list_for_template(
-                    db, template_id=r.template_id,
-                )
-                for f in tpl_files:
-                    files_to_render.append((f, r.template_id))
-
-        for tid in template_ids:
-            if tid not in seen_file_template_ids:
-                seen_file_template_ids.add(tid)
-                tpl_files = await lease_template_file_repo.list_for_template(
-                    db, template_id=tid,
-                )
-                for f in tpl_files:
-                    files_to_render.append((f, tid))
-            for p in await lease_template_placeholder_repo.list_for_template(
-                db, template_id=tid,
-            ):
-                if p.key not in seen_placeholder_keys:
-                    seen_placeholder_keys.add(p.key)
-                    placeholders.append(p)
-
-        # Resolve auto-fillable defaults (applicant / lease / property / user)
-        # for any placeholder with a ``default_source`` set. This populates
-        # ``[ORIGINAL LEASE START DATE]``, ``[PROPERTY ADDRESS]``, etc. on
-        # addendum templates without forcing the host to retype them.
-        applicant, inquiry, property_record, user_record = (
-            await _load_resolution_context(
-                db,
-                lease=lease,
-                organization_id=organization_id,
-                user_id=user_id,
-            )
-        )
-
-        merged_values: dict[str, str] = {
-            k: ("" if v is None else str(v))
-            for k, v in (lease.values or {}).items()
-        }
-        for p in placeholders:
-            existing = merged_values.get(p.key)
-            if existing not in (None, ""):
-                continue
-            if not p.default_source:
-                continue
-            try:
-                resolved, _ = resolve_default_source(
-                    p.default_source,
-                    applicant,
-                    inquiry,
-                    lease=lease,
-                    property_record=property_record,
-                    user_record=user_record,
-                )
-            except (ValueError, AttributeError):
-                logger.warning(
-                    "default_source resolution failed for placeholder %s on lease %s",
-                    p.key, lease_id, exc_info=True,
-                )
-                continue
-            if resolved is not None and resolved != "":
-                merged_values[p.key] = str(resolved)
-
-        # Caller-supplied values win over both existing values and resolved
-        # defaults — they represent the host's explicit choice for this render.
-        if values_override:
-            for k, v in values_override.items():
-                if v is None:
-                    continue
-                merged_values[k] = str(v)
-
-        # Persist merged values to the lease so subsequent regenerations stay
-        # consistent and the Details tab reflects what was used at render time.
-        if merged_values != (lease.values or {}):
-            new_start, new_end = _denormalise_dates(merged_values)
-            fields_to_update: dict[str, Any] = {"values": merged_values}
-            if new_start is not None and lease.starts_on is None:
-                fields_to_update["starts_on"] = new_start
-            if new_end is not None and lease.ends_on is None:
-                fields_to_update["ends_on"] = new_end
-            await signed_lease_repo.update_lease(
-                db,
-                lease_id=lease_id,
-                user_id=user_id,
-                organization_id=organization_id,
-                fields=fields_to_update,
-            )
-
-        # Persist the new template links before rendering (render can be long).
-        # Already-linked templates being regenerated skip this step — their
-        # join row already exists.
-        for order_offset, tid in enumerate(new_template_ids):
-            await signed_lease_template_repo.create(
-                db,
-                lease_id=lease_id,
-                template_id=tid,
-                display_order=max_order + 1 + order_offset,
-            )
-
-        # On regenerate, delete the existing rendered_original attachments
-        # so the new render replaces them. The original imported PDFs (kind=
-        # ``signed_lease`` / ``signed_addendum``) are untouched.
-        rendered_keys_to_delete: list[str] = []
-        if is_regenerate:
-            existing_attachments = await signed_lease_attachment_repo.list_by_lease(
-                db, lease_id,
-            )
-            for att in existing_attachments:
-                if att.kind == "rendered_original":
-                    rendered_keys_to_delete.append(att.storage_key)
-                    await signed_lease_attachment_repo.delete_by_id_scoped_to_lease(
-                        db,
-                        attachment_id=att.id,
-                        lease_id=lease_id,
-                    )
-
-        values_snapshot = dict(merged_values)
-
-    # Best-effort delete of the old rendered objects in MinIO. Run outside the
-    # DB unit-of-work so a storage outage doesn't roll back the DB delete; a
-    # leftover orphan in MinIO is preferable to losing the regenerate.
-    for key in rendered_keys_to_delete:
-        try:
-            storage.delete_file(key)
-        except Exception:  # noqa: BLE001
-            logger.warning(
-                "Failed to delete superseded rendered file %s on regenerate",
-                key, exc_info=True,
-            )
-
-    # Build substitution dict from the merged values (existing + resolved + override).
-    substitutions: dict[str, str] = {
-        k: "" if v is None else str(v) for k, v in values_snapshot.items()
-    }
-    for p in placeholders:
-        if p.input_type == "computed" and p.computed_expr:
-            try:
-                substitutions[p.key] = evaluate(p.computed_expr, values_snapshot)
-            except ComputedExprError as exc:
-                logger.warning(
-                    "Computed placeholder %s failed to evaluate: %s", p.key, exc,
-                )
-                substitutions[p.key] = ""
-
-    # Render each new template's files individually; partial success is fine.
-    now = _dt.datetime.now(_dt.timezone.utc)
-    for f, tid in files_to_render:
-        uploaded_key: str | None = None
-        try:
-            raw = storage.download_file(f.storage_key)
-            if f.content_type == DOCX_MIME:
-                rendered_bytes, _ = render_docx_bytes_to_pdf(raw, substitutions)
-                out_filename = _swap_extension(f.filename, ".pdf")
-                out_ct = "application/pdf"
-            elif f.content_type in ("text/markdown", "text/plain"):
-                text = raw.decode("utf-8", errors="replace")
-                rendered_md_text = render_md(text, substitutions)
-                rendered_bytes = render_md_text_to_pdf_or_keep(rendered_md_text)
-                out_filename = _swap_extension(f.filename, ".pdf")
-                out_ct = "application/pdf"
-            else:
-                rendered_bytes = raw
-                out_filename = f.filename
-                out_ct = f.content_type
-
-            attachment_id = uuid.uuid4()
-            storage_key = f"signed-leases/{lease_id}/{attachment_id}"
-            storage.upload_file(storage_key, rendered_bytes, out_ct)
-            uploaded_key = storage_key
-
-            async with unit_of_work() as db:
-                await signed_lease_attachment_repo.create(
-                    db,
-                    lease_id=lease_id,
-                    storage_key=storage_key,
-                    filename=out_filename,
-                    content_type=out_ct,
-                    size_bytes=len(rendered_bytes),
-                    kind="rendered_original",
-                    uploaded_by_user_id=user_id,
-                    uploaded_at=now,
-                )
-        except Exception:  # noqa: BLE001
-            logger.error(
-                "Failed to render/upload file %s for template %s on lease %s",
-                f.filename,
-                tid,
-                lease_id,
-                exc_info=True,
-            )
-            if uploaded_key is not None:
-                try:
-                    storage.delete_file(uploaded_key)
-                except Exception:  # noqa: BLE001
-                    logger.warning(
-                        "Failed to clean up partially-uploaded file %s", uploaded_key,
-                    )
-
-    return await get_lease(
-        user_id=user_id,
-        organization_id=organization_id,
-        lease_id=lease_id,
-    )
-
-
-async def prefill_addendum_placeholders(
-    *,
-    user_id: uuid.UUID,
-    organization_id: uuid.UUID,
-    lease_id: uuid.UUID,
-    template_ids: list[uuid.UUID],
-) -> SignedLeaseTemplatePrefillResponse:
-    """Compute resolved + unresolved placeholder values for a template + lease pair.
-
-    Walks the placeholders for the requested templates (deduped) and resolves
-    each via ``default_source`` against the parent lease, its applicant /
-    inquiry, the linked property, and the host user. Returns one row per
-    placeholder so the frontend can render a values form that is mostly
-    pre-filled and editable, with empty inputs for genuinely unknown fields.
-
-    Skips placeholders with ``input_type`` of ``signature`` or ``computed`` —
-    those are filled at signing time or evaluated at render time, not by the
-    host in the form.
-    """
-    async with unit_of_work() as db:
-        lease = await signed_lease_repo.get(
-            db,
-            lease_id=lease_id,
-            user_id=user_id,
-            organization_id=organization_id,
-        )
-        if lease is None:
-            raise SignedLeaseNotFoundError(f"Lease {lease_id} not found")
-
-        # Validate every template belongs to the org (cheap, prevents cross-org leak).
-        for tid in template_ids:
-            template = await lease_template_repo.get(
-                db,
-                template_id=tid,
-                user_id=user_id,
-                organization_id=organization_id,
-            )
-            if template is None:
-                raise TemplateNotFoundError(f"Template {tid} not found")
-
-        # Gather + dedupe placeholders across requested templates.
-        placeholders: list = []
-        seen_keys: set[str] = set()
-        for tid in template_ids:
-            for p in await lease_template_placeholder_repo.list_for_template(
-                db, template_id=tid,
-            ):
-                if p.key in seen_keys:
-                    continue
-                seen_keys.add(p.key)
-                placeholders.append(p)
-
-        # Load resolution context once.
-        applicant, inquiry, property_record, user_record = (
-            await _load_resolution_context(
-                db,
-                lease=lease,
-                organization_id=organization_id,
-                user_id=user_id,
-            )
-        )
-
-        existing_values = lease.values or {}
-
-        items: list[SignedLeaseTemplatePrefillItem] = []
-        for p in placeholders:
-            if p.input_type in ("signature", "computed"):
-                continue
-
-            # Existing lease.values wins over default_source so the user's
-            # earlier choices on a generated lease aren't silently overridden.
-            existing = existing_values.get(p.key)
-            if existing not in (None, ""):
-                items.append(
-                    SignedLeaseTemplatePrefillItem(
-                        key=p.key,
-                        display_label=p.display_label,
-                        input_type=p.input_type,
-                        required=p.required,
-                        value=str(existing),
-                        provenance=None,
-                        is_from_existing_values=True,
-                    )
-                )
-                continue
-
-            value: str = ""
-            provenance: str | None = None
-            if p.default_source:
-                try:
-                    resolved, prov = resolve_default_source(
-                        p.default_source,
-                        applicant,
-                        inquiry,
-                        lease=lease,
-                        property_record=property_record,
-                        user_record=user_record,
-                    )
-                    if resolved is not None and resolved != "":
-                        value = str(resolved)
-                        provenance = prov
-                except (ValueError, AttributeError):
-                    logger.warning(
-                        "default_source resolution failed for placeholder %s",
-                        p.key, exc_info=True,
-                    )
-
-            items.append(
-                SignedLeaseTemplatePrefillItem(
-                    key=p.key,
-                    display_label=p.display_label,
-                    input_type=p.input_type,
-                    required=p.required,
-                    value=value,
-                    provenance=provenance,
-                )
-            )
-
-    return SignedLeaseTemplatePrefillResponse(items=items)
-
-
-def should_auto_email_after_generate(
-    *, previous_status: str, auto_email_tenant: bool, last_emailed_to_tenant_at,
-) -> bool:
-    """Pure predicate: should the tenant auto-email fire after generate?
-
-    The four gates that must all be TRUE:
-
-    1. ``previous_status != "generated"`` — Regenerate of an
-       already-generated lease must not re-email. (The user clicked
-       "Regenerate" on purpose; if they want to re-send they use the
-       manual button.)
-    2. ``auto_email_tenant`` — host hasn't opted this lease out of the
-       feature.
-    3. ``last_emailed_to_tenant_at`` is NULL — defensive: even if some
-       prior path (manual? import?) sent a tenant email, don't auto-
-       send again.
-
-    The applicant-has-contact_email check is enforced inside
-    ``lease_email_service.send_lease_to_tenant`` (it skips silently
-    with an info-level event row) — keeping it there means this
-    predicate stays cheap and side-effect-free.
-    """
-    if previous_status == "generated":
-        return False
-    if not auto_email_tenant:
-        return False
-    if last_emailed_to_tenant_at is not None:
-        return False
-    return True
-
-
-async def generate_lease(
-    *,
-    user_id: uuid.UUID,
-    organization_id: uuid.UUID,
-    lease_id: uuid.UUID,
-) -> tuple[SignedLeaseResponse, bool]:
-    """Render the lease and transition status to ``generated``.
-
-    Returns a tuple of ``(detail, should_auto_email)``. The boolean
-    flag is the auto-email gate decision computed BEFORE the status
-    transition (so a Regenerate is correctly identified as
-    ``previous_status="generated"`` and returns False).
-
-    The route handler is responsible for scheduling the auto-email
-    background task; this service stays I/O-pure on the SMTP side.
-    """
-    storage = get_storage()
-    if storage is None:
-        raise StorageNotConfiguredError("Object storage is not configured")
-
-    async with unit_of_work() as db:
-        lease = await signed_lease_repo.get(
-            db,
-            lease_id=lease_id,
-            user_id=user_id,
-            organization_id=organization_id,
-        )
-        if lease is None:
-            raise SignedLeaseNotFoundError(f"Lease {lease_id} not found")
-        # Snapshot the gate inputs BEFORE we mutate the lease.
-        previous_status = lease.status
-        auto_email_tenant = lease.auto_email_tenant
-        last_emailed_to_tenant_at = lease.last_emailed_to_tenant_at
-
-        # Load every template linked to this lease, then load each template's
-        # files + placeholders. Files are processed in template-display-order
-        # so the rendered attachments appear in the host's pick order.
-        join_rows = await signed_lease_template_repo.list_for_lease(
-            db, lease_id=lease_id,
-        )
-        files: list = []
-        placeholders: list = []
-        seen_placeholder_keys: set[str] = set()
-        for jr in join_rows:
-            tpl_files = await lease_template_file_repo.list_for_template(
-                db, template_id=jr.template_id,
-            )
-            files.extend(tpl_files)
-            tpl_placeholders = await lease_template_placeholder_repo.list_for_template(
-                db, template_id=jr.template_id,
-            )
-            # Dedupe placeholders across templates — first definition wins
-            # (matches the merge rule used at draft creation time).
-            for p in tpl_placeholders:
-                if p.key not in seen_placeholder_keys:
-                    seen_placeholder_keys.add(p.key)
-                    placeholders.append(p)
-
-    # Build the substitution dict — start from raw values, then evaluate
-    # computed placeholders (which can reference other keys).
-    substitutions: dict[str, str] = {}
-    for key, value in (lease.values or {}).items():
-        substitutions[key] = "" if value is None else str(value)
-    for p in placeholders:
-        if p.input_type == "computed" and p.computed_expr:
-            try:
-                substitutions[p.key] = evaluate(p.computed_expr, lease.values or {})
-            except ComputedExprError as exc:
-                logger.warning(
-                    "Computed placeholder %s failed to evaluate: %s", p.key, exc,
-                )
-                substitutions[p.key] = ""
-
-    # Render each file and upload.
-    uploaded: list[tuple[str, str, str, int]] = []  # (storage_key, filename, content_type, size)
-    try:
-        for f in files:
-            raw = storage.download_file(f.storage_key)
-            if f.content_type == DOCX_MIME:
-                # DOCX templates render to PDF: substitute placeholders via
-                # python-docx, convert to markdown via mammoth, then to PDF
-                # via reportlab. Single output format (PDF) for every
-                # generated lease — easier for hosts and tenants to handle.
-                rendered_bytes, _ = render_docx_bytes_to_pdf(raw, substitutions)
-                out_filename = _swap_extension(f.filename, ".pdf")
-                out_ct = "application/pdf"
-            elif f.content_type in ("text/markdown", "text/plain"):
-                text = raw.decode("utf-8", errors="replace")
-                rendered_md_text = render_md(text, substitutions)
-                rendered_bytes = render_md_text_to_pdf_or_keep(rendered_md_text)
-                out_filename = _swap_extension(f.filename, ".pdf")
-                out_ct = "application/pdf"
-            else:
-                # Pass-through (defensive — upload allowlist shouldn't allow
-                # other content types).
-                rendered_bytes = raw
-                out_filename = f.filename
-                out_ct = f.content_type
-
-            attachment_id = uuid.uuid4()
-            storage_key = f"signed-leases/{lease_id}/{attachment_id}"
-            storage.upload_file(storage_key, rendered_bytes, out_ct)
-            uploaded.append((storage_key, out_filename, out_ct, len(rendered_bytes)))
-
-        # Persist attachment rows + transition lease to "generated".
-        async with unit_of_work() as db:
-            now = _dt.datetime.now(_dt.timezone.utc)
-            for storage_key, out_filename, out_ct, size in uploaded:
-                await signed_lease_attachment_repo.create(
-                    db,
-                    lease_id=lease_id,
-                    storage_key=storage_key,
-                    filename=out_filename,
-                    content_type=out_ct,
-                    size_bytes=size,
-                    kind="rendered_original",
-                    uploaded_by_user_id=user_id,
-                    uploaded_at=now,
-                )
-            await signed_lease_repo.update_lease(
-                db,
-                lease_id=lease_id,
-                user_id=user_id,
-                organization_id=organization_id,
-                fields={"status": "generated", "generated_at": now},
-            )
-    except Exception:
-        # Best-effort cleanup of just-uploaded objects on failure.
-        for storage_key, *_ in uploaded:
-            try:
-                storage.delete_file(storage_key)
-            except Exception:  # noqa: BLE001
-                logger.warning("Failed to clean up rendered file %s", storage_key)
-        raise
-
-    detail = await get_lease(
-        user_id=user_id,
-        organization_id=organization_id,
-        lease_id=lease_id,
-    )
-    should_auto_email = should_auto_email_after_generate(
-        previous_status=previous_status,
-        auto_email_tenant=auto_email_tenant,
-        last_emailed_to_tenant_at=last_emailed_to_tenant_at,
-    )
-    return detail, should_auto_email
-
-
-def render_md_text_to_pdf_or_keep(rendered_md: str) -> bytes:
-    """Render rendered markdown to PDF bytes (low-fidelity, reportlab-based)."""
-    return render_pdf_from_text(rendered_md)
-
-
-def _ensure_suffix(filename: str, suffix: str) -> str:
-    if filename.lower().endswith(suffix):
-        return filename
-    base = filename.rsplit(".", 1)[0] if "." in filename else filename
-    return f"{base}{suffix}"
-
-
-def _swap_extension(filename: str, suffix: str) -> str:
-    base = filename.rsplit(".", 1)[0] if "." in filename else filename
-    return f"{base}{suffix}"
-
-
-# ---------------------------------------------------------------------------
-# Import signed lease (externally-signed PDFs — no template required)
-# ---------------------------------------------------------------------------
-
-# Attachment-kind heuristic for import uploads.
-# The FIRST file is always signed_lease.  Subsequent files check the filename
-# for "move" + "in" → move_in_inspection, or "move" + "out" →
-# move_out_inspection.  Everything else is signed_addendum.
-# Heuristic is deliberately conservative — false negatives are cheaper than
-# false positives (wrong kind mislabels the file in the UI, but it's editable).
-def _infer_attachment_kind(filename: str, position: int) -> str:
-    if position == 0:
-        return "signed_lease"
-    lower = filename.lower()
-    if "move" in lower and "out" in lower:
-        return "move_out_inspection"
-    if "move" in lower and "in" in lower:
-        return "move_in_inspection"
-    return "signed_addendum"
-
-
-class ApplicantNotFoundError(LookupError):
-    pass
-
-
-class ListingNotFoundError(LookupError):
-    pass
-
-
-async def import_signed_lease(
-    *,
-    user_id: uuid.UUID,
-    organization_id: uuid.UUID,
-    applicant_id: uuid.UUID,
-    listing_id: uuid.UUID | None,
-    starts_on: _dt.date | None,
-    ends_on: _dt.date | None,
-    notes: str | None,
-    status: str,
-    files: list[tuple[bytes, str, str | None]],  # (content, filename, declared_ct)
-) -> SignedLeaseResponse:
-    """Create an imported signed lease from externally-signed PDFs.
-
-    Unlike ``create_lease``, this path does NOT require a template. The lease
-    is created with ``kind='imported'``, no template links, and
-    ``signed_at=now()`` since by definition the documents are already signed.
-
-    ``files`` is an ordered list of ``(content_bytes, filename, content_type)``
-    tuples. The first file becomes ``kind=signed_lease``; subsequent files use
-    the ``_infer_attachment_kind`` heuristic.
-    """
-    from app.core.config import settings as _settings
-
-    storage = get_storage()
-    if storage is None:
-        raise StorageNotConfiguredError("Object storage is not configured")
-
-    # Validate all files before touching the database.
-    processed: list[tuple[bytes, str, str]] = []  # (content, filename, ct)
-    for content, filename, declared_ct in files:
-        if len(content) > _settings.max_blackout_attachment_size_bytes:
-            max_mb = _settings.max_blackout_attachment_size_bytes // (1024 * 1024)
-            raise AttachmentTooLargeError(f"File '{filename}' exceeds {max_mb}MB limit")
-        ct = _resolve_content_type(content, filename, declared_ct)
-        if ct is None:
-            raise AttachmentTypeRejectedError(
-                f"Unsupported file type for '{filename}'. "
-                "Allowed: pdf, docx, jpg, png, webp",
-            )
-        # EXIF-strip images to remove GPS metadata.
-        if ct in ("image/jpeg", "image/png", "image/webp"):
-            content = _exif_strip_image(content, ct)
-        processed.append((content, filename, ct))
-
-    # Validate tenant scoping.
-    async with unit_of_work() as db:
-        applicant = await applicant_repo.get(
-            db,
-            applicant_id=applicant_id,
-            organization_id=organization_id,
-            user_id=user_id,
-        )
-        if applicant is None:
-            raise ApplicantNotFoundError(f"Applicant {applicant_id} not found")
-
-        if listing_id is not None:
-            listing = await listing_repo.get_by_id(
-                db,
-                listing_id=listing_id,
-                organization_id=organization_id,
-            )
-            if listing is None:
-                raise ListingNotFoundError(f"Listing {listing_id} not found")
-
-        now = _dt.datetime.now(_dt.timezone.utc)
-        lease = await signed_lease_repo.create(
-            db,
-            user_id=user_id,
-            organization_id=organization_id,
-            applicant_id=applicant_id,
-            listing_id=listing_id,
-            values={},
-            starts_on=starts_on,
-            ends_on=ends_on,
-            status=status,
-            kind="imported",
-        )
-        # Stamp signed_at — imported leases are signed by definition.
-        await signed_lease_repo.update_lease(
-            db,
-            lease_id=lease.id,
-            user_id=user_id,
-            organization_id=organization_id,
-            fields={"signed_at": now, "notes": notes},
-        )
-        lease_id = lease.id
-
-    # Upload files and persist attachment rows.
-    uploaded: list[str] = []  # storage keys for rollback on failure
-    try:
-        async with unit_of_work() as db:
-            now = _dt.datetime.now(_dt.timezone.utc)
-            for position, (content, filename, ct) in enumerate(processed):
-                attachment_id = uuid.uuid4()
-                storage_key = f"signed-leases/{lease_id}/{attachment_id}"
-                storage.upload_file(storage_key, content, ct)
-                uploaded.append(storage_key)
-                kind = _infer_attachment_kind(filename, position)
-                await signed_lease_attachment_repo.create(
-                    db,
-                    lease_id=lease_id,
-                    storage_key=storage_key,
-                    filename=filename or f"attachment-{attachment_id.hex}",
-                    content_type=ct,
-                    size_bytes=len(content),
-                    kind=kind,
-                    uploaded_by_user_id=user_id,
-                    uploaded_at=now,
-                )
-    except Exception:
-        for storage_key in uploaded:
-            try:
-                storage.delete_file(storage_key)
-            except Exception:  # noqa: BLE001
-                logger.warning("Failed to clean up import attachment %s", storage_key)
-        raise
-
-    return await get_lease(
-        user_id=user_id,
-        organization_id=organization_id,
-        lease_id=lease_id,
-    )
-
-
-def _resolve_content_type(
-    content: bytes, filename: str, declared: str | None,
-) -> str | None:
-    """Return a validated MIME type or None if not in the allowlist."""
-    if declared and declared in ALLOWED_ATTACHMENT_MIME_TYPES:
-        return declared
-    lower = filename.lower()
-    ext_map = {
-        ".pdf": "application/pdf",
-        ".docx": "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-        ".jpg": "image/jpeg",
-        ".jpeg": "image/jpeg",
-        ".png": "image/png",
-        ".webp": "image/webp",
-    }
-    for ext, ct in ext_map.items():
-        if lower.endswith(ext):
-            return ct
-    return None
-
-
-def _exif_strip_image(content: bytes, content_type: str) -> bytes:
-    """EXIF-strip an image via Pillow. Returns cleaned bytes."""
-    from app.services.storage.image_processor import process_image, ImageRejected
-    try:
-        result = process_image(content, declared_content_type=content_type)
-        return result.content
-    except ImageRejected:
-        # If Pillow can't decode it, let it pass through — the content-type
-        # check already validated the header bytes.
-        return content
-
-
-# ---------------------------------------------------------------------------
-# Soft-delete
-# ---------------------------------------------------------------------------
-
-async def soft_delete_lease(
-    *,
-    user_id: uuid.UUID,
-    organization_id: uuid.UUID,
-    lease_id: uuid.UUID,
-) -> None:
-    async with unit_of_work() as db:
-        lease = await signed_lease_repo.get(
-            db,
-            lease_id=lease_id,
-            user_id=user_id,
-            organization_id=organization_id,
-        )
-        if lease is None:
-            raise SignedLeaseNotFoundError(f"Lease {lease_id} not found")
-        await signed_lease_repo.soft_delete(
-            db,
-            lease_id=lease_id,
-            user_id=user_id,
-            organization_id=organization_id,
-        )
-
-
-# ---------------------------------------------------------------------------
 # Attachments
-# ---------------------------------------------------------------------------
+from app.services.leases.lease_attachment_service import (
+    delete_attachment,
+    list_attachments,
+    update_attachment_kind,
+    update_attachment_signing_state,
+    upload_attachment,
+)
 
-async def upload_attachment(
-    *,
-    user_id: uuid.UUID,
-    organization_id: uuid.UUID,
-    lease_id: uuid.UUID,
-    content: bytes,
-    filename: str,
-    declared_content_type: str | None,
-    kind: str,
-) -> SignedLeaseAttachmentResponse:
-    storage = get_storage()
-    if storage is None:
-        raise StorageNotConfiguredError("Object storage is not configured")
-
-    from app.core.config import settings as _settings
-
-    if kind not in LEASE_ATTACHMENT_KINDS:
-        raise AttachmentTypeRejectedError(f"Invalid kind: {kind}")
-
-    # Tenant scope.
-    async with unit_of_work() as db:
-        lease = await signed_lease_repo.get(
-            db,
-            lease_id=lease_id,
-            user_id=user_id,
-            organization_id=organization_id,
-        )
-        if lease is None:
-            raise SignedLeaseNotFoundError(f"Lease {lease_id} not found")
-
-    if len(content) > _settings.max_blackout_attachment_size_bytes:
-        max_mb = _settings.max_blackout_attachment_size_bytes // (1024 * 1024)
-        raise AttachmentTooLargeError(f"File exceeds {max_mb}MB limit")
-
-    # Resolve content type — fall back to extension when octet-stream.
-    if declared_content_type and declared_content_type in ALLOWED_ATTACHMENT_MIME_TYPES:
-        ct = declared_content_type
-    else:
-        lower = filename.lower()
-        if lower.endswith(".pdf"):
-            ct = "application/pdf"
-        elif lower.endswith(".docx"):
-            ct = DOCX_MIME
-        elif lower.endswith(".jpg") or lower.endswith(".jpeg"):
-            ct = "image/jpeg"
-        elif lower.endswith(".png"):
-            ct = "image/png"
-        elif lower.endswith(".webp"):
-            ct = "image/webp"
-        else:
-            raise AttachmentTypeRejectedError(
-                f"Unsupported attachment type. Allowed: pdf, docx, jpg, png, webp",
-            )
-    if ct not in ALLOWED_ATTACHMENT_MIME_TYPES:
-        raise AttachmentTypeRejectedError(
-            f"Unsupported attachment type ({ct}). Allowed: pdf, docx, jpg, png, webp",
-        )
-
-    attachment_id = uuid.uuid4()
-    storage_key = f"signed-leases/{lease_id}/{attachment_id}"
-    storage.upload_file(storage_key, content, ct)
-
-    try:
-        async with unit_of_work() as db:
-            row = await signed_lease_attachment_repo.create(
-                db,
-                lease_id=lease_id,
-                storage_key=storage_key,
-                filename=filename or f"attachment-{attachment_id.hex}",
-                content_type=ct,
-                size_bytes=len(content),
-                kind=kind,
-                uploaded_by_user_id=user_id,
-                uploaded_at=_dt.datetime.now(_dt.timezone.utc),
-            )
-            response = SignedLeaseAttachmentResponse.model_validate(row)
-    except Exception:
-        try:
-            storage.delete_file(storage_key)
-        except Exception:  # noqa: BLE001
-            logger.warning("Failed to clean up orphan lease attachment %s", storage_key)
-        raise
-
-    return attach_presigned_urls_to_attachments([response])[0]
-
-
-async def list_attachments(
-    *,
-    user_id: uuid.UUID,
-    organization_id: uuid.UUID,
-    lease_id: uuid.UUID,
-) -> list[SignedLeaseAttachmentResponse]:
-    async with unit_of_work() as db:
-        lease = await signed_lease_repo.get(
-            db,
-            lease_id=lease_id,
-            user_id=user_id,
-            organization_id=organization_id,
-        )
-        if lease is None:
-            raise SignedLeaseNotFoundError(f"Lease {lease_id} not found")
-        rows = await signed_lease_attachment_repo.list_by_lease(db, lease_id)
-    return _attachment_responses(rows)
-
-
-async def update_attachment_signing_state(
-    *,
-    user_id: uuid.UUID,
-    organization_id: uuid.UUID,
-    lease_id: uuid.UUID,
-    attachment_id: uuid.UUID,
-    signing_fields: dict[str, _dt.datetime | None],
-) -> SignedLeaseAttachmentResponse:
-    """Set / clear signing-state timestamps on a lease attachment.
-
-    ``signing_fields`` is the subset of ``{"signed_by_tenant_at",
-    "signed_by_landlord_at"}`` the host explicitly set on the request
-    body — keys absent from the body are left untouched. ``None`` values
-    explicitly clear that party's signature. Field-name validation is
-    enforced at the schema layer (Pydantic ``extra="forbid"``); this
-    function trusts the keys it receives.
-    """
-    async with unit_of_work() as db:
-        lease = await signed_lease_repo.get(
-            db,
-            lease_id=lease_id,
-            user_id=user_id,
-            organization_id=organization_id,
-        )
-        if lease is None:
-            raise SignedLeaseNotFoundError(f"Lease {lease_id} not found")
-
-        row = await signed_lease_attachment_repo.update_signing_state_scoped_to_lease(
-            db,
-            attachment_id=attachment_id,
-            lease_id=lease_id,
-            fields=signing_fields,
-        )
-        if row is None:
-            raise AttachmentNotFoundError(f"Attachment {attachment_id} not found")
-
-        response = SignedLeaseAttachmentResponse.model_validate(row)
-
-    return attach_presigned_urls_to_attachments([response])[0]
-
-
-async def update_attachment_kind(
-    *,
-    user_id: uuid.UUID,
-    organization_id: uuid.UUID,
-    lease_id: uuid.UUID,
-    attachment_id: uuid.UUID,
-    kind: str,
-) -> SignedLeaseAttachmentResponse:
-    """Change the kind of an existing attachment.
-
-    Validates the kind is in ``LEASE_ATTACHMENT_KINDS``, then applies a
-    composite-WHERE update (attachment_id + lease_id) to prevent IDOR.
-    """
-    if kind not in LEASE_ATTACHMENT_KINDS:
-        raise InvalidAttachmentKindError(f"Invalid kind: {kind}")
-
-    async with unit_of_work() as db:
-        # Tenant scope — 404 if lease doesn't belong to this org/user.
-        lease = await signed_lease_repo.get(
-            db,
-            lease_id=lease_id,
-            user_id=user_id,
-            organization_id=organization_id,
-        )
-        if lease is None:
-            raise SignedLeaseNotFoundError(f"Lease {lease_id} not found")
-
-        row = await signed_lease_attachment_repo.update_kind_scoped_to_lease(
-            db,
-            attachment_id=attachment_id,
-            lease_id=lease_id,
-            kind=kind,
-        )
-        if row is None:
-            raise AttachmentNotFoundError(f"Attachment {attachment_id} not found")
-
-        response = SignedLeaseAttachmentResponse.model_validate(row)
-
-    return attach_presigned_urls_to_attachments([response])[0]
-
-
-async def delete_attachment(
-    *,
-    user_id: uuid.UUID,
-    organization_id: uuid.UUID,
-    lease_id: uuid.UUID,
-    attachment_id: uuid.UUID,
-) -> None:
-    async with unit_of_work() as db:
-        lease = await signed_lease_repo.get(
-            db,
-            lease_id=lease_id,
-            user_id=user_id,
-            organization_id=organization_id,
-        )
-        if lease is None:
-            raise SignedLeaseNotFoundError(f"Lease {lease_id} not found")
-        deleted = await signed_lease_attachment_repo.delete_by_id_scoped_to_lease(
-            db, attachment_id, lease_id,
-        )
-        if deleted is None:
-            raise AttachmentNotFoundError(f"Attachment {attachment_id} not found")
-        storage_key = deleted.storage_key
-
-    storage = get_storage()
-    if storage is not None:
-        try:
-            storage.delete_file(storage_key)
-        except Exception:  # noqa: BLE001
-            logger.warning(
-                "Failed to delete lease attachment object %s",
-                storage_key, exc_info=True,
-            )
+__all__ = [
+    # Exceptions
+    "AttachmentNotFoundError",
+    "AttachmentTooLargeError",
+    "AttachmentTypeRejectedError",
+    "ALLOWED_ATTACHMENT_MIME_TYPES",
+    "CannotEditValuesError",
+    "InvalidAttachmentKindError",
+    "InvalidStatusTransitionError",
+    "MissingRequiredValuesError",
+    "SignedLeaseNotFoundError",
+    "StorageNotConfiguredError",
+    "ImportedLeaseTemplateError",
+    "TemplatesAlreadyLinkedError",
+    "ApplicantNotFoundError",
+    "ListingNotFoundError",
+    # Lifecycle
+    "create_lease",
+    "get_lease",
+    "list_leases",
+    "soft_delete_lease",
+    "update_lease",
+    # PDF generation
+    "add_templates_and_generate",
+    "generate_lease",
+    "prefill_addendum_placeholders",
+    "render_md_text_to_pdf_or_keep",
+    "should_auto_email_after_generate",
+    # Import
+    "import_signed_lease",
+    "infer_kind_from_filename",
+    "infer_kinds_for_files",
+    # Attachments
+    "delete_attachment",
+    "list_attachments",
+    "update_attachment_kind",
+    "update_attachment_signing_state",
+    "upload_attachment",
+]
