+++ apps/mybookkeeper/backend/tests/test_lease_attachment_kind.py
@@ -18,7 +18,7 @@ from app.core.context import RequestContext
 from app.core.permissions import current_org_member, require_write_access
 from app.main import app
 from app.models.organization.organization_member import OrgRole
-from app.services.leases.signed_lease_service import (
+from app.services.leases.lease_import_service import (
     infer_kind_from_filename,
     infer_kinds_for_files,
 )
